Moved stopAreaView into /nodeView folder
