+++ src/components/sidebar/nodeView/stopAreaView/stopAreaView.scss
@@ -1,7 +1,7 @@
-@import '../../shared/styles/common';
-@import '../../shared/styles/sidebar';
-@import '../../shared/styles/flexbox';
-@import '../../shared/styles/form';
+@import '../../../shared/styles/common';
+@import '../../../shared/styles/sidebar';
+@import '../../../shared/styles/flexbox';
+@import '../../../shared/styles/form';
 
 .stopAreaView {
     width: $sidebarWide;
